source first

+++ src/test/update.sh
@@ -5,21 +5,25 @@
 #
 # shellcheck disable=SC1090,SC1091
 
-pushd "$1" || exit 1
+pushd "$1"/src || exit 1
+source bkg.sh
+popd || exit 1
 
-if [ -d index ]; then
-    [ ! -d index.bak ] || rm -rf index.bak
-    mv index index.bak
+if git ls-remote --exit-code origin index &>/dev/null; then
+    if [ -d index ]; then
+        [ ! -d index.bak ] || rm -rf index.bak
+        mv index index.bak
+    fi
+
+    git fetch origin index
+    git worktree add index index
+    pushd index || exit 1
+    git reset --hard origin/index
+    popd || exit 1
 fi
 
-git fetch origin index
-git worktree add index index
-pushd index || exit 1
-git reset --hard origin/index
-popd || exit 1
 [ -f index/.env ] && \cp index/.env src/env.env || echo "No .env file found"
 pushd src || exit 1
-source bkg.sh
 main "${@:2}"
 
 check_json() {
